% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_movie.R
\name{get_movie}
\alias{get_movie}
\title{A Movie Get Function}
\usage{
get_movie(key_word, id_num)
}
\arguments{
\item{A}{number the user chooses}
}
\value{
A list of results of a specific movie
}
\description{
This function allows you to get the movie info.
}
