% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imdb_search.R
\name{imdb_search}
\alias{imdb_search}
\title{This internal function returns the dataframe of movie.}
\usage{
imdb_search(imdb_id)
}
\arguments{
\item{imdb_id}{search id}
}
\value{
A dataframe that contains the list of search results.
Dataframe contains Title, Type, Year, and Director.
}
\description{
This internal function returns the dataframe of movie.
}
