% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plot.R
\name{read_plot}
\alias{read_plot}
\title{This function returns the plot of Movies and TV shows based on search word(s) by user.}
\usage{
read_plot(search_text, type = "")
}
\arguments{
\item{search_text}{search text i.e., word(s) to look for in the Title.}

\item{type}{expected type of the result can be - movie, series, episode, game}
}
\value{
A dataframe that contains the list of search results.
Dataframe contains Title, Type, Year, and Plot.
}
\description{
This function returns the plot of Movies and TV shows based on search word(s) by user.
}
\examples{
read_plot('Guardians')
read_plot('Guardians Galaxy', type = 'movie')
}
