% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_movie_info.R
\name{search_movie}
\alias{search_movie}
\title{A Movie Search Function}
\usage{
search_movie(keyWord)
}
\arguments{
\item{A}{string of an English word}
}
\value{
A list of search results
}
\description{
This function allows you to search movie by using a key word.
}
