% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_movie.R
\name{select_movie}
\alias{select_movie}
\title{A Movie Select Function}
\usage{
select_movie(js_lst, id_num)
}
\arguments{
\item{A}{number the user chooses}
}
\value{
A list of results of a specific movie
}
\description{
This function allows you to select the movie by choosing the corresponding number.
}
